+++ modules/nlu/src/backend/predict-pipeline.ts
@@ -18,7 +18,7 @@ export type ExactMatchResult = (sdk.MLToolkit.SVM.Prediction & { extractor: 'exa
 export type Predictors = TrainArtefacts & {
   ctx_classifier: sdk.MLToolkit.SVM.Predictor
   intent_classifier_per_ctx: _.Dictionary<sdk.MLToolkit.SVM.Predictor>
-  oos_classifier: sdk.MLToolkit.SVM.Predictor
+  oos_classifier_per_ctx: _.Dictionary<sdk.MLToolkit.SVM.Predictor>
   kmeans: sdk.MLToolkit.KMeans.KmeansResult
   slot_tagger: SlotTagger // TODO replace this by MlToolkit.CRF.Tagger
   pattern_entities: PatternEntity[]
@@ -46,7 +46,7 @@ export type PredictStep = {
     elected?: E1IntentPred // only to comply with E1
     ambiguous?: boolean
   }
-  oos_predictions?: sdk.MLToolkit.SVM.Prediction
+  oos_predictions?: _.Dictionary<number>
   slot_predictions_per_intent?: _.Dictionary<SlotExtractionResult[]>
 }
 
@@ -61,7 +61,7 @@ type E1IntentPred = {
 
 const DEFAULT_CTX = 'global'
 const NONE_INTENT = 'none'
-const OOS_AS_NONE_TRESH = 0.3
+const OOS_AS_NONE_TRESH = 0.4
 const LOW_INTENT_CONFIDENCE_TRESH = 0.4
 
 async function DetectLanguage(
@@ -237,12 +237,15 @@ async function predictIntent(input: PredictStep, predictors: Predictors): Promis
   const ctxToPredict = input.ctx_predictions.map(p => p.label)
   const predictions = (
     await Promise.map(ctxToPredict, async ctx => {
+      let preds: sdk.MLToolkit.SVM.Prediction[] = []
+
       const predictor = predictors.intent_classifier_per_ctx[ctx]
-      if (!predictor) {
-        return
+      if (predictor) {
+        const features = [...input.utterance.sentenceEmbedding, input.utterance.tokens.length] // TODO: extract this logic 'getIntentFeatures()' in a place for intent featurizing
+        const tmp = await predictor.predict(features)
+        preds.push(...tmp)
       }
-      const features = [...input.utterance.sentenceEmbedding, input.utterance.tokens.length]
-      let preds = await predictor.predict(features)
+
       const exactPred = findExactIntentForCtx(predictors.exact_match_index, input.utterance, ctx)
       if (exactPred) {
         const idxToRemove = preds.findIndex(p => p.label === exactPred.label)
@@ -252,7 +255,13 @@ async function predictIntent(input: PredictStep, predictors: Predictors): Promis
 
       if (input.alternateUtterance) {
         const alternateFeats = [...input.alternateUtterance.sentenceEmbedding, input.alternateUtterance.tokens.length]
-        const alternatePreds = await predictor.predict(alternateFeats)
+
+        const alternatePreds: sdk.MLToolkit.SVM.Prediction[] = []
+        if (predictor) {
+          const tmp = await predictor.predict(alternateFeats)
+          alternatePreds.push(...tmp)
+        }
+
         const exactPred = findExactIntentForCtx(predictors.exact_match_index, input.alternateUtterance, ctx)
         if (exactPred) {
           const idxToRemove = alternatePreds.findIndex(p => p.label === exactPred.label)
@@ -260,14 +269,13 @@ async function predictIntent(input: PredictStep, predictors: Predictors): Promis
           alternatePreds.unshift(exactPred)
         }
 
-        // we might want to do this in intent election intead or in NDU
-        if ((alternatePreds && alternatePreds[0]?.confidence) ?? 0 >= preds[0].confidence) {
-          // mean
+        if (
+          (alternatePreds && alternatePreds.filter(p => p.label !== NONE_INTENT)[0]?.confidence) ??
+          0 >= preds.filter(p => p.label !== NONE_INTENT)[0].confidence
+        ) {
           preds = _.chain([...alternatePreds, ...preds])
             .groupBy('label')
-            .mapValues(gr => _.meanBy(gr, 'confidence'))
-            .toPairs()
-            .map(([label, confidence]) => ({ label, confidence }))
+            .map(preds => _.maxBy(preds, 'confidence'))
             .value()
         }
       }
@@ -316,12 +324,12 @@ function electIntent(input: PredictStep): PredictStep {
     .flatMap(({ label: ctx, confidence: ctxConf }) => {
       const intentPreds = _.chain(input.intent_predictions.per_ctx[ctx] || [])
         .thru(preds => {
-          if (input.oos_predictions?.confidence > OOS_AS_NONE_TRESH) {
+          if (input.oos_predictions[ctx] >= OOS_AS_NONE_TRESH) {
             return [
               ...preds,
               {
                 label: NONE_INTENT,
-                confidence: input.oos_predictions?.confidence ?? 1,
+                confidence: input.oos_predictions[ctx],
                 context: ctx,
                 l0Confidence: ctxConf
               }
@@ -332,13 +340,18 @@ function electIntent(input: PredictStep): PredictStep {
         })
         .map(p => ({ ...p, confidence: _.round(p.confidence, 2) }))
         .orderBy('confidence', 'desc')
-        .value()
-      if (intentPreds[0].confidence === 1 || intentPreds.length === 1) {
+        .value() as (sdk.MLToolkit.SVM.Prediction & { context: string })[]
+      if (intentPreds[0]?.confidence === 1 || intentPreds.length === 1) {
         return [{ label: intentPreds[0].label, l0Confidence: ctxConf, context: ctx, confidence: 1 }]
-      } // are we sure theres always at least two intents ? otherwise down there it may crash
+      }
 
+      const noneIntent = { label: NONE_INTENT, context: ctx, confidence: 1 }
       if (predictionsReallyConfused(intentPreds)) {
-        intentPreds.unshift({ label: NONE_INTENT, context: ctx, confidence: 1 })
+        intentPreds.unshift(noneIntent)
+      }
+
+      if (intentPreds.length <= 1) {
+        return noneIntent
       }
 
       const lnstd = math.std(intentPreds.filter(x => x.confidence !== 0).map(x => Math.log(x.confidence))) // because we want a lognormal distribution
@@ -368,12 +381,12 @@ function electIntent(input: PredictStep): PredictStep {
     predictions.length &&
     predictions[0].name !== NONE_INTENT &&
     predictions[0].confidence < LOW_INTENT_CONFIDENCE_TRESH &&
-    input.oos_predictions?.confidence > OOS_AS_NONE_TRESH
+    input.oos_predictions[ctx] > OOS_AS_NONE_TRESH
   if (!predictions.length || shouldConsiderOOS) {
     predictions = _.orderBy(
       [
         ...predictions.filter(p => p.name !== NONE_INTENT),
-        { name: NONE_INTENT, context: ctx, confidence: input.oos_predictions?.confidence ?? 1 }
+        { name: NONE_INTENT, context: ctx, confidence: input.oos_predictions[ctx] || 1 }
       ],
       'confidence'
     )
@@ -384,18 +397,29 @@ function electIntent(input: PredictStep): PredictStep {
   })
 }
 
-async function predictOutOfScope(input: PredictStep, predictors: Predictors, tools: Tools): Promise<PredictStep> {
-  if (!isPOSAvailable(input.languageCode) || !predictors.oos_classifier) {
-    return input
+async function predictOutOfScope(input: PredictStep, predictors: Predictors): Promise<PredictStep> {
+  const oos_predictions = {}
+  if (!isPOSAvailable(input.languageCode) || _.isEmpty(predictors.oos_classifier_per_ctx)) {
+    return {
+      ...input,
+      oos_predictions: Object.keys(predictors.contexts).reduce((preds, ctx) => ({ ...preds, [ctx]: 0 }), {})
+    }
   }
+
   const utt = input.alternateUtterance || input.utterance
-  const feats = getUtteranceFeatures(utt)
-  const preds = await predictors.oos_classifier.predict(feats)
-  const confidence = _.sumBy(
-    preds.filter(p => p.label.startsWith('out')),
-    'confidence'
-  )
-  const oos_predictions = { label: 'out', confidence }
+  const feats = getUtteranceFeatures(utt, predictors.vocabVectors)
+  for (const ctx of predictors.contexts) {
+    try {
+      const preds = await predictors.oos_classifier_per_ctx[ctx].predict(feats)
+      oos_predictions[ctx] = _.chain(preds)
+        .filter(p => p.label.startsWith('out'))
+        .maxBy('confidence')
+        .get('confidence', 0)
+        .value()
+    } catch (err) {
+      oos_predictions[ctx] = 0
+    }
+  }
 
   return {
     ...input,
@@ -440,6 +464,7 @@ async function extractSlots(input: PredictStep, predictors: Predictors): Promise
 }
 
 function MapStepToOutput(step: PredictStep, startTime: number): PredictOutput {
+  // legacy pre-ndu
   const entities = step.utterance.entities.map(
     e =>
       ({
@@ -450,6 +475,7 @@ function MapStepToOutput(step: PredictStep, startTime: number): PredictOutput {
           value: e.value
         },
         meta: {
+          sensitive: e.sensitive,
           confidence: e.confidence,
           end: e.endPos,
           source: e.metadata.source,
@@ -457,6 +483,7 @@ function MapStepToOutput(step: PredictStep, startTime: number): PredictOutput {
         }
       } as sdk.NLU.Entity)
   )
+  // legacy pre-ndu
   const slots = step.utterance.slots.reduce((slots, s) => {
     return {
       ...slots,
@@ -471,51 +498,48 @@ function MapStepToOutput(step: PredictStep, startTime: number): PredictOutput {
     }
   }, {} as sdk.NLU.SlotCollection)
 
-  const predictions = step.ctx_predictions?.reduce(
-    (preds, { label, confidence }) => {
-      return {
-        ...preds,
-        [label]: {
-          confidence: confidence,
-          intents: step.intent_predictions.per_ctx[label].map(i => ({
-            ...i,
-            slots: (step.slot_predictions_per_intent[i.label] || []).reduce((slots, s) => {
-              if (slots[s.slot.name] && slots[s.slot.name].confidence > s.slot.confidence) {
-                // we keep only the most confident slots
-                return slots
-              }
+  const slotsCollectionReducer = (slots: sdk.NLU.SlotCollection, s: SlotExtractionResult): sdk.NLU.SlotCollection => {
+    if (slots[s.slot.name] && slots[s.slot.name].confidence > s.slot.confidence) {
+      // we keep only the most confident slots
+      return slots
+    }
 
-              return {
-                ...slots,
-                [s.slot.name]: {
-                  start: s.start,
-                  end: s.end,
-                  confidence: s.slot.confidence,
-                  name: s.slot.name,
-                  source: s.slot.source,
-                  value: s.slot.value
-                } as sdk.NLU.Slot
-              }
-            }, {} as sdk.NLU.SlotCollection)
-          }))
-        }
-      }
-    },
-    {
-      oos: {
-        intents: [
-          {
-            label: NONE_INTENT,
-            confidence: 1 // this will be be computed as
-          }
-        ],
-        confidence: step.oos_predictions?.confidence ?? 0
+    return {
+      ...slots,
+      [s.slot.name]: {
+        start: s.start,
+        end: s.end,
+        confidence: s.slot.confidence,
+        name: s.slot.name,
+        source: s.slot.source,
+        value: s.slot.value
+      } as sdk.NLU.Slot
+    }
+  }
+
+  const predictions: sdk.NLU.Predictions = step.ctx_predictions.reduce((preds, current) => {
+    const { label, confidence } = current
+
+    const intentPred = step.intent_predictions.per_ctx[label]
+    const intents = !intentPred
+      ? []
+      : intentPred.map(i => ({
+          ...i,
+          slots: (step.slot_predictions_per_intent[i.label] || []).reduce(slotsCollectionReducer, {})
+        }))
+
+    return {
+      ...preds,
+      [label]: {
+        confidence: confidence,
+        oos: step.oos_predictions[label] || 0,
+        intents
       }
     }
-  )
+  }, {})
 
   return {
-    ambiguous: step.intent_predictions.ambiguous,
+    ambiguous: step.intent_predictions.ambiguous, // legacy pre-ndu
     detectedLanguage: step.detectedLanguage,
     entities,
     errored: false,
@@ -524,16 +548,15 @@ function MapStepToOutput(step: PredictStep, startTime: number): PredictOutput {
       .orderBy(x => x[1].confidence, 'desc')
       .fromPairs()
       .value(),
-    includedContexts: step.includedContexts,
-    intent: step.intent_predictions.elected,
-    intents: step.intent_predictions.combined,
+    includedContexts: step.includedContexts, // legacy pre-ndu
+    intent: step.intent_predictions.elected, // legacy pre-ndu
+    intents: step.intent_predictions.combined, // legacy pre-ndu
     language: step.languageCode,
-    slots,
+    slots, // legacy pre-ndu
     ms: Date.now() - startTime
   }
 }
 
-// TODO move this in exact match module
 export function findExactIntentForCtx(
   exactMatchIndex: ExactMatchIndex,
   utterance: Utterance,
@@ -566,7 +589,7 @@ export const Predict = async (
 
     stepOutput = await makePredictionUtterance(stepOutput, predictors, tools)
     stepOutput = await extractEntities(stepOutput, predictors, tools)
-    stepOutput = await predictOutOfScope(stepOutput, predictors, tools)
+    stepOutput = await predictOutOfScope(stepOutput, predictors)
     stepOutput = await predictContext(stepOutput, predictors)
     stepOutput = await predictIntent(stepOutput, predictors)
     stepOutput = electIntent(stepOutput)