+++ modules/qna/src/backend/qna.d.ts
@@ -20,6 +20,9 @@ export interface QnaEntry {
 
 export interface QnaItem {
   id: string
+  isNew?: boolean
+  key?: string
+  saveError?: string
   data: QnaEntry
 }