+++ src/bp/sdk/botpress.d.ts
@@ -25,6 +25,29 @@ declare module 'botpress/sdk' {
     ): Promise<T>
   }
 
+  export interface Incident {
+    id: string
+    ruleName: string
+    hostName: string
+    startTime: Date
+    endTime?: Date
+    triggerValue: number
+  }
+
+  export type StrategyUser = {
+    id?: number
+    password?: string
+    salt?: string
+  } & UserInfo
+
+  export interface UserInfo {
+    email: string
+    strategy: string
+    createdOn?: string
+    updatedOn?: string
+    attributes: any
+  }
+
   export type KnexExtended = Knex & KnexExtension
 
   /**
@@ -218,6 +241,7 @@ declare module 'botpress/sdk' {
 
   export interface FlowGeneratorMetadata {
     botId: string
+    isOneFlow?: boolean
   }
 
   export interface ModulePluginEntry {
@@ -480,6 +504,7 @@ declare module 'botpress/sdk' {
     }
 
     export interface EntityMeta {
+      sensitive: boolean
       confidence: number
       provider: string
       source: string
@@ -503,7 +528,8 @@ declare module 'botpress/sdk' {
     export interface Predictions {
       [context: string]: {
         confidence: number
-        intents: { label: string; confidence: number }[]
+        oos: number
+        intents: { label: string; confidence: number; slots: SlotCollection }[]
       }
     }
   }
@@ -713,6 +739,8 @@ declare module 'botpress/sdk' {
       bot: any
       /** Used internally by Botpress to keep the user's current location and upcoming instructions */
       context: DialogContext
+      /** This variable points to the currently active workflow */
+      workflow: WorkflowHistory
       /**
        * EXPERIMENTAL
        * This includes all the flow/nodes which were traversed for the current event
@@ -737,6 +765,8 @@ declare module 'botpress/sdk' {
       node: string
       /** When a jump point is used, it will be removed from the list on the next transition */
       used?: boolean
+      /** When true, the node targeted by this jump point will be executed from the start (instead of only transitions) */
+      executeNode?: boolean
     }
 
     export interface DialogContext {
@@ -763,16 +793,20 @@ declare module 'botpress/sdk' {
       lastMessages: DialogTurnHistory[]
       nluContexts?: NluContext[]
       nduContext?: NduContext
-      lastWorkflows: WorkflowHistory[]
+      workflows: {
+        [name: string]: WorkflowHistory
+      }
+      currentWorkflow?: string
       // Prevent warnings when using the code editor with custom properties
       [anyKey: string]: any
     }
 
     export interface WorkflowHistory {
-      workflow: string
       eventId: string
+      parent?: string
+      /** Only one workflow can be active at a time, when a child workflow is active, the parent will be pending */
+      status: 'active' | 'pending' | 'completed'
       success?: boolean
-      active?: boolean
     }
 
     export type StoredEvent = {
@@ -1284,7 +1318,7 @@ declare module 'botpress/sdk' {
     triggers: { conditions: DecisionTriggerCondition[] }[]
   }
 
-  export type SkillFlowNode = Partial<ListenNode> & Pick<Required<ListenNode>, 'name'>
+  export type SkillFlowNode = Partial<ListenNode> & Pick<Required<ListenNode>, 'name'> & Partial<TriggerNode>
 
   /**
    * Node Transitions are all the possible outcomes when a user's interaction on a node is completed. The possible destinations